More logic for Annotation support: 1. Parsing various types of DB Annotations to PSDFKit Annotations 2. Calculation of bounding boxes. 3. Color and blend mode calculations 4. Extraction of various info needed for annotations from RItem

+++ app/src/main/java/org/zotero/android/api/mappers/UserResponseMapper.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class UserResponseMapper @Inject constructor() {
 
     fun fromJson(json: JsonObject): UserResponse {
-        val id = json["id"].asInt
+        val id = json["id"].asLong
         val name = json["name"].asString
         val username = json["username"].asString
         return UserResponse(id = id, name = name, username = username)
